pyproject.toml
[tool.poetry]
name = "outcome-utils"
version = "0.1.0"
description = "A collection of python utils."
authors = ["Douglas Willcocks <douglas@outcome.co>"]
license = "ISC"
packages = [
  { include = "outcome", from = "src" }
]
homepage = "https://github.com/outcome-co/utils-py"
repository = "https://github.com/outcome-co/utils-py"
documentation = "https://github.com/outcome-co/utils-py#README"

[tool.poetry.dependencies]
python = "^3.7"

[tool.poetry.dev-dependencies]
outcome-devkit-py = "^0.6.0"
pytest-asyncio = "^0.12.0"

[tool.coverage.run]
branch = true
data_file = 'coverage/data'
source = ['src']
omit = [
   "test/*"
]
parallel = true

[tool.coverage.report]
fail_under = 100
skip_empty = true

[tool.coverage.html]
show_contexts = true
directory = 'coverage/html'

[tool.black]
line-length = 130
include = '\.pyi?$'
# This stops black from rewriting quotes
skip-string-normalization = true
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | \.cache
)/
'''

[tool.isort]
skip_glob = "*/.cache/**/*"

[tool.commitizen]
name = "cz_conventional_commits"
version = "0.1.0"
tag_format = "v$version"
bump_message = "chore(version): $current_version → $new_version"
version_files = ["pyproject.toml:version"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

